+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.0.0
 
+- Contracts: complete stake requested event  ([#30](https://github.com/OpenSTFoundation/openst-tokens/pull/30))
+- Contracts: add requestStake  ([#30](https://github.com/OpenSTFoundation/openst-tokens/pull/30))
 - Contracts: add value token to constructor ([#25](https://github.com/OpenSTFoundation/openst-tokens/pull/25))
 - Contracts: inherit EIP20 interface ([#25](https://github.com/OpenSTFoundation/openst-tokens/pull/25))
 - Contracts: import EIP20 interface ([#25](https://github.com/OpenSTFoundation/openst-tokens/pull/25))